+++ src/app/[locale]/jobs/form/(routes)/vehicle-information/data/vehicle-data.ts
