+++ src/app/[locale]/jobs/layout.tsx
@@ -1,0 +1,18 @@
+import Header from './components/header';
+
+function JobsLayout({
+  children,
+  params: { locale }
+}: Readonly<{
+  children: React.ReactNode;
+  params: { locale: string };
+}>) {
+  return (
+    <div>
+      <Header locale={locale} />
+      <main>{children}</main>
+    </div>
+  );
+}
+
+export default JobsLayout;
