--- src/app/layout.tsx
@@ -1,29 +1,0 @@
-import type { Metadata } from 'next';
-import { Poppins } from 'next/font/google';
-import './globals.css';
-
-import ReactQueryProvider from '@/components/ui/react-query-provider';
-
-const poppins = Poppins({
-  subsets: ['latin'],
-  weight: ['400', '500', '600']
-});
-
-export const metadata: Metadata = {
-  title: 'ABC Emergency',
-  description: 'Generated by create next app'
-};
-
-export default function RootLayout({
-  children
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body className={`${poppins.className} antialiased`}>
-        <ReactQueryProvider>{children}</ReactQueryProvider>
-      </body>
-    </html>
-  );
-}
