--- src/app/page.tsx
@@ -1,7 +1,0 @@
-'use server';
-
-import { redirect } from 'next/navigation';
-
-export default async function Home() {
-  redirect('/jobs');
-}
